# docs/openapi.yaml
openapi: 3.1.0
info:
  title: NexusTale API
  version: "1.0.0"
  description: >
    REST API for NexusTale — a sci-fi/fantasy novel-writing platform with
    Git-backed versioning, world wiki, and AI co-author support.
servers:
  - url: /api/v1
    description: Current API prefix (served by nginx reverse-proxy in all environments)
security: []  # default unauthenticated; individual operations override via bearerAuth
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ── Reusable error shapes ───────────────────────────────────────────────────
  responses:
    Unauthorized:
      description: Missing or invalid JWT.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    BadRequest:
      description: Validation or binding error.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        message: { type: string }
        detail:  { type: string }
    MessageResponse:
      type: object
      required: [message]
      properties:
        message: { type: string }

    # ── Auth ──────────────────────────────────────────────────────────────────
...
    CreateProjectRequest:
      type: object
      required: [title]
      properties:
        title:       { type: string, minLength: 1, maxLength: 200 }
        description: { type: string }
        genres:      { type: array, items: { type: string } }
    UpdateProjectRequest:
      type: object
      properties:
        title:       { type: string, minLength: 1, maxLength: 200 }
        description: { type: string }
    ProjectResponse:
      type: object
      required: [id, owner_id, title, description, genres, archived, phase, auto_tag_enabled, created_at, updated_at]
      properties:
        id:               { type: string, format: uuid }
        owner_id:         { type: string, format: uuid }
        title:            { type: string }
        description:      { type: string }
        genres:           { type: array, items: { type: string } }
        archived:         { type: boolean }
        phase:            { type: string }
        auto_tag_enabled: { type: boolean }
        created_at:       { type: string, format: date-time }
        updated_at:       { type: string, format: date-time }
...
    EntityType:
      type: string
      enum: [character, location, faction, item, concept, lore]
...
    EntityResponse:
      type: object
      required: [id, project_id, type, name, summary, attributes, created_at, updated_at]
      properties:
        id:               { type: string, format: uuid }
        project_id:       { type: string, format: uuid }
        parent_entity_id: { type: string, format: uuid, nullable: true }
        type:             { $ref: '#/components/schemas/EntityType' }
        name:             { type: string }
        summary:          { type: string }
        attributes:       { type: object, additionalProperties: { type: string } }
        image_url:        { type: string, nullable: true, description: "Presigned MinIO URL valid for ~4 hours; absent when no image is stored." }
        created_at:       { type: string, format: date-time }
        updated_at:       { type: string, format: date-time }
        rename_cascade_available: { type: boolean, description: "Set when a name change has ≥1 scene mention. Triggers cascade banner in UI." }
        occurrence_count:         { type: integer, description: "Total non-suppressed mentions of the old name." }
        old_name:                 { type: string, description: "Previous name before the update; only set when rename_cascade_available is true." }

    # ── Wiki — Relationships ──────────────────────────────────────────────────
...
    PromptCategory:
      type: string
      enum: [style, voice, workshop, system]
    CreatePromptRequest:
      type: object
      required: [name, content]
      properties:
        name:     { type: string }
        content:  { type: string }
        category: { $ref: '#/components/schemas/PromptCategory' }
...
    PromptResponse:
      type: object
      required: [id, project_id, name, content, category, created_at, updated_at]
      properties:
        id:         { type: string, format: uuid }
        project_id: { type: string, format: uuid }
        name:       { type: string }
        content:    { type: string }
        category:   { $ref: '#/components/schemas/PromptCategory' }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }

    # ── Export ────────────────────────────────────────────────────────────────
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      summary: List all projects owned by the authenticated user
      tags: [projects]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: List of projects.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ProjectResponse' }
        "401": { $ref: '#/components/responses/Unauthorized' }
    post:
      operationId: createProject
      summary: Create a new project
      tags: [projects]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateProjectRequest' }
      responses:
        "201":
          description: Project created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProjectResponse' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
  /projects/{projectId}:
    parameters:
      - name: projectId
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      operationId: getProject
      summary: Get a single project
      tags: [projects]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Project details.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProjectResponse' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "404": { $ref: '#/components/responses/NotFound' }
    patch:
      operationId: updateProject
      summary: Update project title or description
      tags: [projects]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateProjectRequest' }
      responses:
        "200":
          description: Updated project.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProjectResponse' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "404": { $ref: '#/components/responses/NotFound' }
    delete:
      operationId: deleteProject
      summary: Delete a project
      tags: [projects]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Project deleted.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MessageResponse' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "404": { $ref: '#/components/responses/NotFound' }
...
  /projects/{projectId}/wiki/entities/{entityId}/image:
    parameters:
      - name: projectId
        in: path
        required: true
        schema: { type: string, format: uuid }
      - name: entityId
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      operationId: uploadEntityImage
      summary: Upload or replace an entity portrait image
      tags: [wiki]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [image]
              properties:
                image:
                  type: string
                  format: binary
      responses:
        "200":
          description: Entity with updated image_url.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EntityResponse' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "404": { $ref: '#/components/responses/NotFound' }
    delete:
      operationId: deleteEntityImage
      summary: Remove an entity's portrait image
      tags: [wiki]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Entity with image_url cleared.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EntityResponse' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "404": { $ref: '#/components/responses/NotFound' }

  # ── Wiki — Relationships ──────────────────────────────────────────────────────
...
  /projects/{projectId}/prompts:
    parameters:
      - { name: projectId, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      operationId: listPrompts
      summary: List writing-style prompts for a project
      tags: [prompts]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: List of prompts.
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/PromptResponse' } }
        "401": { $ref: '#/components/responses/Unauthorized' }
    post:
      operationId: createPrompt
      summary: Create a writing-style prompt
      tags: [prompts]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreatePromptRequest' }
      responses:
        "201":
          description: Created prompt.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PromptResponse' }
        "401": { $ref: '#/components/responses/Unauthorized' }
...